FlowControl intro
device and host memory cache are separate now
