+++ nd4j-backends/nd4j-backend-impls/nd4j-cuda-7.5/src/main/java/org/nd4j/jita/allocator/flow/impl/AsynchronousFlowController.java
@@ -1,0 +1,7 @@
+package org.nd4j.jita.allocator.flow.impl;
+
+/**
+ * @author raver119@gmail.com
+ */
+public class AsynchronousFlowController {
+}
