+++ nd4j-backends/nd4j-backend-impls/nd4j-cuda-7.5/src/main/java/org/nd4j/jita/memory/impl/CudaCachingZeroProvider.java
@@ -20,41 +20,41 @@ import java.util.concurrent.atomic.AtomicLong;
 
 
 /**
- * This is MemoryProvider implementation, that adds cache for memory reuse purposes.
+ * This is MemoryProvider implementation, that adds cache for memory reuse purposes. Only host memory is cached for future reuse.
+ *
  * If some memory chunk gets released via allocator, it'll be probably saved for future reused within same JVM process.
  *
  * @author raver119@gmail.com
  */
-public class CudaCachingProvider extends CudaDirectProvider implements MemoryProvider {
-    private static Logger log = LoggerFactory.getLogger(CudaCachingProvider.class);
+public class CudaCachingZeroProvider extends CudaDirectProvider implements MemoryProvider {
+    private static Logger log = LoggerFactory.getLogger(CudaCachingZeroProvider.class);
 
-    private volatile ConcurrentHashMap<AllocationShape, CacheHolder> zeroCache = new ConcurrentHashMap<>();
-    private volatile ConcurrentHashMap<Integer, ConcurrentHashMap<AllocationShape, CacheHolder>> deviceCache = new ConcurrentHashMap<>();
+    protected volatile ConcurrentHashMap<AllocationShape, CacheHolder> zeroCache = new ConcurrentHashMap<>();
 
-    private final AtomicLong cacheHit = new AtomicLong(0);
-    private final AtomicLong cacheMiss = new AtomicLong(0);
+    protected final AtomicLong cacheHit = new AtomicLong(0);
+    protected final AtomicLong cacheMiss = new AtomicLong(0);
 
     private final AtomicLong allocRequests = new AtomicLong(0);
 
     private final AtomicLong zeroCachedAmount = new AtomicLong(0);
-    private final AtomicLong deviceCachedAmount = new AtomicLong(0);
 
-    private final Semaphore singleLock = new Semaphore(1);
+
+    protected final Semaphore singleLock = new Semaphore(1);
 
     // we don't cache allocations greater then this value
-    private final long MAX_SINGLE_ALLOCATION = 1000000;
+    protected final long MAX_SINGLE_ALLOCATION = 1000000;
 
     // maximum cached size of memory
-    private final long MAX_CACHED_MEMORY;
+    protected final long MAX_CACHED_MEMORY;
 
     // memory chunks below this threshold will be guaranteed regardless of number of cache entries
     // that especially covers all possible variations of shapeInfoDataBuffers in all possible cases
-    private final long FORCED_CACHE_THRESHOLD = 96;
+    protected final long FORCED_CACHE_THRESHOLD = 96;
 
     //  number of preallocation entries for each yet-unknown shape
-    private final int PREALLOCATION_LIMIT = 50;
+    protected final int PREALLOCATION_LIMIT = 50;
 
-    public CudaCachingProvider() {
+    public CudaCachingZeroProvider() {
         MAX_CACHED_MEMORY = Runtime.getRuntime().maxMemory() / 2;
     }
 
@@ -104,61 +104,10 @@ public class CudaCachingProvider extends CudaDirectProvider implements MemoryPro
             return super.malloc(shape, point, location);
         }
 
-        if (location == AllocationStatus.DEVICE && reqMemory < MAX_SINGLE_ALLOCATION) {
-            ensureDeviceCacheHolder(point.getDeviceId(), shape);
-
-            CacheHolder cache = deviceCache.get(point.getDeviceId()).get(shape);
-            if (cache != null) {
-                Pointer pointer = cache.poll();
-                if (pointer != null) {
-                    cacheHit.incrementAndGet();
-
-                    deviceCachedAmount.addAndGet(-1 * reqMemory);
-
-                    PointersPair pair = new PointersPair();
-                    pair.setDevicePointer(new CudaPointer(pointer.getNativePointer()));
-                    pair.setHostPointer(new CudaPointer(pointer.getNativePointer()));
-
-                    point.setAllocationStatus(AllocationStatus.DEVICE);
-                    return pair;
-                }
-            }
-            cacheMiss.incrementAndGet();
-            return super.malloc(shape, point, location);
-        }
-
         return super.malloc(shape, point, location);
     }
 
-    protected  void ensureDeviceCacheHolder(Integer deviceId, AllocationShape shape) {
-        if (!deviceCache.containsKey(deviceId)) {
-            try {
-                singleLock.acquire();
 
-                if (!deviceCache.containsKey(deviceId)) {
-                    deviceCache.put(deviceId, new ConcurrentHashMap<AllocationShape, CacheHolder>());
-                }
-            } catch (Exception e) {
-                throw new RuntimeException(e);
-            } finally {
-                singleLock.release();
-            }
-        }
-
-        if (!deviceCache.get(deviceId).containsKey(shape)) {
-            try {
-                singleLock.acquire();
-
-                if (!deviceCache.get(deviceId).containsKey(shape)) {
-                    deviceCache.get(deviceId).put(shape, new CacheHolder(shape));
-                }
-            } catch (Exception e) {
-
-            } finally {
-                singleLock.release();
-            }
-        }
-    }
 
     protected void ensureCacheHolder(AllocationShape shape) {
         if (!zeroCache.containsKey(shape)) {
@@ -186,38 +135,7 @@ public class CudaCachingProvider extends CudaDirectProvider implements MemoryPro
     @Override
     public void free(AllocationPoint point) {
         if (point.getAllocationStatus() == AllocationStatus.DEVICE) {
-            AllocationShape shape = point.getShape();
-            long reqMemory = AllocationUtils.getRequiredMemory(shape);
-            // we don't cache too big objects
-
-            if (reqMemory > MAX_SINGLE_ALLOCATION || deviceCachedAmount.get() >= MAX_CACHED_MEMORY) {
-                super.free(point);
-                return;
-            }
-
-            ensureDeviceCacheHolder(point.getDeviceId(), shape);
-
-            /*
-                Now we should decide if this object can be cached or not
-             */
-            CacheHolder cache = deviceCache.get(point.getDeviceId()).get(shape);
-
-            // memory chunks < threshold will be cached no matter what
-            if (reqMemory <= FORCED_CACHE_THRESHOLD) {
-                cache.put(new Pointer(point.getDevicePointer().address()));
-            } else {
-                long cacheEntries = cache.size();
-                long cacheHeight = deviceCache.get(point.getDeviceId()).size();
-
-                // total memory allocated within this bucket
-                long cacheDepth = cacheEntries * reqMemory;
-
-                if (cacheDepth < MAX_CACHED_MEMORY / cacheHeight) {
-                    cache.put(new Pointer(point.getDevicePointer().address()));
-                } else {
-                    super.free(point);
-                }
-            }
+            super.free(point);
         } else {
             AllocationShape shape = point.getShape();
             long reqMemory = AllocationUtils.getRequiredMemory(shape);
@@ -317,10 +235,10 @@ public class CudaCachingProvider extends CudaDirectProvider implements MemoryPro
             for (int i = 0; i < target; i ++) {
                 AllocationPoint point = new AllocationPoint();
 
-                PointersPair pair = CudaCachingProvider.super.malloc(shape, point, this.location);
+                PointersPair pair = CudaCachingZeroProvider.super.malloc(shape, point, this.location);
                 if (this.location == AllocationStatus.HOST) {
                     Pointer pointer = new Pointer(pair.getHostPointer().address());
-                    CudaCachingProvider.this.zeroCache.get(shape).put(pointer);
+                    CudaCachingZeroProvider.this.zeroCache.get(shape).put(pointer);
                 }
             }
         }
