+++ nd4j-backends/nd4j-backend-impls/nd4j-cuda-7.5/src/main/java/org/nd4j/linalg/jcublas/buffer/BaseCudaDataBuffer.java
@@ -357,6 +357,7 @@ public abstract class BaseCudaDataBuffer extends BaseDataBuffer implements JCuda
     @Override
     public void setData(int[] data) {
         set(data, data.length, 0, 0);
+        allocator.synchronizeHostData(this);
     }
 
     @Override
